refactor: split native API into focused route files, remove dead gateway code
Decompose 757-line api/native/inference.rb into:
- api/native/helpers.rb — shared helpers (ClientToolMethods, parse_request_body,
  validate_*, json_response, json_error, etc.)
- api/native/chat.rb — POST /api/llm/chat
- api/native/providers.rb — GET /api/llm/providers, GET /api/llm/providers/:name
- api/native/inference.rb — POST /api/llm/inference (slimmed down)

Remove all gateway_available? references and gateway branches from:
- api/native/inference.rb (definition + 3 call sites)
- fleet/dispatcher.rb (gateway transport message fallback)

Update api.rb to register all sub-routes via Native::* modules.
Add Routes = API constant alias for backward compatibility.
Provider routes now use Router health tracker and Call::Registry
instead of the removed gateway ProviderStats dependency.
